Set baseline to 3.6 everywhere
