+++ instrumentation/vertx/vertx-kafka-client-3.6/testing/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+  id("otel.java-conventions")
+}
+
+dependencies {
+  implementation(project(":testing-common"))
+  implementation("org.testcontainers:kafka")
+
+  compileOnly("io.vertx:vertx-kafka-client:3.6.0")
+  // vertx-codegen is needed for Xlint's annotation checking
+  compileOnly("io.vertx:vertx-codegen:3.6.0")
+}
