+++ settings.gradle.kts
@@ -429,8 +429,8 @@ include(":instrumentation:vaadin-14.2:testing")
 include(":instrumentation:vertx:vertx-http-client:vertx-http-client-3.0:javaagent")
 include(":instrumentation:vertx:vertx-http-client:vertx-http-client-4.0:javaagent")
 include(":instrumentation:vertx:vertx-http-client:vertx-http-client-common:javaagent")
-include(":instrumentation:vertx:vertx-kafka-client-3.5:javaagent")
-include(":instrumentation:vertx:vertx-kafka-client-3.5:testing")
+include(":instrumentation:vertx:vertx-kafka-client-3.6:javaagent")
+include(":instrumentation:vertx:vertx-kafka-client-3.6:testing")
 include(":instrumentation:vertx:vertx-rx-java-3.5:javaagent")
 include(":instrumentation:vertx:vertx-web-3.0:javaagent")
 include(":instrumentation:vertx:vertx-web-3.0:testing")
